Make NativeExecutionContext authoritative and harden the public native path.
Unify residency/events under one Rust execution context, add NativeStreamingStore
and batched Compute region invocation, include the PyO3 crate in workspace gates,
and robustify cancel/thread/residency failure modes for concurrent forwards.

+++ crates/streamcompiler-virtual-backend/src/backend.rs
@@ -218,7 +218,7 @@ impl Backend for VirtualBackend {
             supports_p2p: self.config.supports_p2p,
             supports_async_compute: true,
             supports_ordered_streams: true,
-            max_streams: 8.max(self.config.max_copy_engines as u32 + 2),
+            max_streams: 8.max(self.config.max_copy_engines + 2),
             device_memory_bytes: self.config.memory_bytes,
             simulated: true,
         }
